added zone factor logic to Air and Ground zone

+++ src/Week4/TimeZone/TimeZoneBehaviour.java
@@ -3,5 +3,5 @@ package Week4.TimeZone;
 import Week4.Address;
 
 public interface TimeZoneBehaviour {
-    public int daysToDeliver(Address origin, Address destination);
+    public double daysToDeliver(Address origin, Address destination);
 }
